requirements.txt
python-pptx==0.6.21
openai
click==8.1.7
python-dotenv==1.0.0
llama-parse
pydantic
